submit code

+++ papers/dense_lu/debug/lu.cpp
@@ -194,10 +194,10 @@ bool lu_factorize_core(double* A, int* piv, int n) {
         }
 
         // Eliminate below diagonal
-        double akk = M(A, k, k, n);
+        float akk = M(A, k, k, n);
         for (int i = k + 1; i < n; ++i) {
             M(A, i, k, n) /= akk;                            // multiplier → L
-            double lik = M(A, i, k, n);
+            float lik = M(A, i, k, n);
             for (int j = k + 1; j < n; ++j)
                 M(A, i, j, n) -= lik * M(A, k, j, n);       // Schur complement
         }
